New implementations and Some changes (#59)
* HolyC implementation commit

Implemented using TempleOS API.

* Fix rewriting errors commit

* PowerShell implementation commit

* Objective-C implementation commit

Rewrote a function in a library project that was suddenly found to search for information in keys for OS X

* Rename catalog commit

":" is special character inside Microsoft Windows for determination the file-streams. NTFS driver recognize "1С:Предприятие" like filestream -> Git for Windows throws I/O panic.

I hope it normally -- directory renaming :3

* add newlines

* I/O format changed commit

I found mistakes here.

* reimplementation of Objective-C program commit

* fixed binary-operation issue commit

not 13 * 2. 26.
It was a headache, but i solved it.

* fix newlines

---------

Co-authored-by: Kira Malinova <llathasa@outlook.com>